refactor: DateText.vueをatoms/へ
